trace duplicate registrations (#694)
Summary:
While using ClassyVision, I was getting a bit confused as to how duplicate registrations were occurring. It turned out to be a consequence of the way a test harness of a downstream project did imports. I have a sense others may struggle with understanding such error messages about duplicates, too. In this change, I add full tracebacks of the original registration to the errors, which is very verbose but potentially useful.

Pull Request resolved: https://github.com/facebookresearch/ClassyVision/pull/694

Reviewed By: mannatsingh

Differential Revision: D26074256

Pulled By: bottler

fbshipit-source-id: a00098f5ed6c3370c2d4833a1e073de9186927a8